+++ protected/humhub/components/Controller.php
@@ -22,6 +22,11 @@ class Controller extends \yii\web\Controller
 {
 
     /**
+     * @event \yii\base\Event an event raised on init a controller.
+     */
+    const EVENT_INIT = 'create';
+
+    /**
      * @var null|string the name of the sub layout to be applied to this controller's views.
      * This property mainly affects the behavior of [[render()]].
      */
@@ -41,6 +46,15 @@ class Controller extends \yii\web\Controller
      * @var boolean append page title 
      */
     public $prependActionTitles = true;
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+        $this->trigger(self::EVENT_INIT);
+    }
 
     /**
      * @inheritdoc
@@ -92,7 +106,7 @@ class Controller extends \yii\web\Controller
         if (\Yii::$app->request->method != 'POST') {
             throw new \yii\web\HttpException(405, Yii::t('ContentModule.controllers_ContentController', 'Invalid request method!'));
         }
-        
+
         return true;
     }
 
@@ -143,7 +157,7 @@ class Controller extends \yii\web\Controller
                 $this->getView()->pageTitle = $this->pageTitle;
             }
 
-            if(!Yii::$app->request->isAjax || Yii::$app->request->isPjax) {
+            if (!Yii::$app->request->isAjax || Yii::$app->request->isPjax) {
                 $this->setJsViewStatus();
             }
 
@@ -223,7 +237,7 @@ class Controller extends \yii\web\Controller
         $modluleId = (Yii::$app->controller->module) ? Yii::$app->controller->module->id : '';
         $this->view->registerJs('humhub.modules.ui.view.setState("' . $modluleId . '", "' . Yii::$app->controller->id . '", "' . Yii::$app->controller->action->id . '");', \yii\web\View::POS_BEGIN);
 
-        if(Yii::$app->request->isPjax) {
+        if (Yii::$app->request->isPjax) {
             \humhub\widgets\TopMenu::setViewState();
         }
     }