+++ protected/humhub/docs/CHANGELOG.md
@@ -74,6 +74,7 @@ HumHub Change Log
 - Fix: All LDAP Users have been disabled and not reenabled by hourly cronjob if ldap server not reachable.
 - Enh: Cleanup authentication layout file 
 - Fix: Console image converter memory limit allocation
+- Enh: Added new controller init event
 
 1.2.0-beta.2 (February 24, 2017)
 --------------------------------